Make DirectSoundRenderThread::GetPlayPositionMs return double
Also make all internal storage/arithmetic related to it use doubles. The main motivation for this is that it allows us to work around some of the issues identified in #59, as it avoids 64-bit integer arithmetic altogether. doubles have 52 significand bits, so by keeping track of bytes read by DirectSound, it should be accurate up to 7 million hours or something (as opposed to about 6 hours with 32 bit integers, and even that is a probably a bit overkill actually, except for perhaps a musicdisk and perhaps some extra bits for headroom).

+++ WaveSabrePlayerLib/src/DirectSoundRenderThread.cpp
@@ -26,7 +26,7 @@ namespace WaveSabrePlayerLib
 		WaitForSingleObject(thread, INFINITE);
 	}
 
-	int DirectSoundRenderThread::GetPlayPositionMs()
+	double DirectSoundRenderThread::GetPlayPositionMs()
 	{
 		if (!buffer)
 			return 0;
@@ -35,7 +35,7 @@ namespace WaveSabrePlayerLib
 		buffer->GetCurrentPosition((LPDWORD)&playCursorPos, 0);
 
 		int currentOldPlayCursorPos;
-		long long currentBytesRendered;
+		double currentBytesRendered;
 
 		{
 			auto playPositionCriticalSectionGuard = playPositionCriticalSection.Enter();
@@ -44,12 +44,12 @@ namespace WaveSabrePlayerLib
 			currentBytesRendered = bytesRendered;
 		}
 
-		long long totalBytesRead = playCursorPos - currentOldPlayCursorPos;
+		double totalBytesRead = playCursorPos - currentOldPlayCursorPos;
 		if (totalBytesRead < 0)
 			totalBytesRead += bufferSizeBytes;
 		totalBytesRead += currentBytesRendered;
 
-		return (int)(totalBytesRead / SongRenderer::BlockAlign * 1000 / sampleRate);
+		return totalBytesRead / SongRenderer::BlockAlign * 1000 / sampleRate;
 	}
 
 	DWORD WINAPI DirectSoundRenderThread::threadProc(LPVOID lpParameter)
